add colliders
