+++ Assets/Scripts/DCLRuntime/ComponentHandlers/MeshColliderHandler.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 32562fa15b67465c96d9129ebaf03a42
+timeCreated: 1682678022
