--- Assets/Scripts/DCLRuntime/ComponentManager.ComponentHandler.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 70cdadc0ecd24194b61f16bac209e977
-timeCreated: 1682665543
