+++ Assets/Scripts/DCLRuntime/ComponentManager.cs
@@ -1,11 +1,14 @@
 using System;
 using DCL.CRDT;
 using DCL.ECS7;
+using DCL.ECSComponents;
+using DCLRuntime.ComponentHandlers;
+using UnityEngine;
 using Object = UnityEngine.Object;
 
 namespace DCLRuntime
 {
-    public partial class ComponentManager : IDisposable
+    public class ComponentManager : IDisposable
     {
         private readonly SceneRoot _sceneRoot = SceneRoot.Create();
 
@@ -16,6 +19,28 @@ namespace DCLRuntime
             {
                 var entity = _sceneRoot.GetCreateEntity(message.EntityId);
                 HandleComponent(entity, message.ComponentId, message.Timestamp, message.Data);
+            }
+        }
+        
+        private void HandleComponent(GameObject entity, int componentId, int timestamp, object data)
+        {
+            switch (componentId)
+            {
+                case ComponentID.TRANSFORM:
+                    var transformData = TransformData.FromData((byte[])data);
+                    transformData.ApplyOn(entity.transform);
+                    break;
+                case ComponentID.MESH_RENDERER:
+                    var meshRendererData = ProtoSerialization.Deserialize<PBMeshRenderer>(data);
+                    meshRendererData.Apply(entity);
+                    break;
+                case ComponentID.MESH_COLLIDER:
+                    var meshColliderData = ProtoSerialization.Deserialize<PBMeshCollider>(data);
+                    meshColliderData.Apply(entity);
+                    break;
+                default:
+                    Debug.LogError($"not supported component {componentId}");
+                    break;
             }
         }
